Beginning renaming "wrappers", which was a terrible name given how much they did, to "managers"
